pyproject.toml
[project]
name = "refinance-tool"
version = "0.1.0"
description = "Agentic AI tool that helps users determine if they should refinance their mortgage."
authors = [
    {name = "lukeschwenke"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "langgraph (>=1.0.2,<2.0.0)",
    "langchain-openai (>=1.0.2,<2.0.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "tavily-python (>=0.7.12,<0.8.0)",
    "langchain-mcp-adapters (>=0.1.12,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "refinance_tool"
packages = [
    { include = "core", from = "src" }
]

[tool.pytest.ini_options]
# (optional) so tests can also import without installing
pythonpath = ["src"]
markers = ["integration: tests that hit live endpoints"]